Simplify timing point display on timeline
